added Buttons to component file
